Move inference tests with suggestions
